proyecto_clase/proyecto_23319/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.2
sqlparse==0.4.3
